agrego test de stats

+++ src/main/java/com/example/challengemutanteml/StatsADN.java
@@ -1,15 +1,15 @@
 package com.example.challengemutanteml;
 
 public class StatsADN {
-    public int count_mutant_dna;
-    public int count_human_dna;
+    public double count_mutant_dna;
+    public double count_human_dna;
     public double ratio;
 
-    public int getCount_mutant_dna() {
+    public double getCount_mutant_dna() {
         return count_mutant_dna;
     }
 
-    public int getCount_human_dna() {
+    public double getCount_human_dna() {
         return count_human_dna;
     }
